LightBar detections

+++ include/tools.hpp
@@ -16,7 +16,7 @@ public:
     static void drawLightbars(cv::Mat &image, const Lightbar_Pair &inputPairs, const cv::Scalar &color, int thickness);
     static cv::Mat enhanceContrast(const cv::Mat &inputFrame);
     static std::vector<std::vector<cv::Point>> getContours(cv::Mat &inputFrame,double minAreaThreshold);
-    static cv::Mat adjustBrightness(const cv::Mat &inputFrame, int beta);
+    static cv::Mat adjustBrightness(const cv::Mat &inputFrame, double beta);
     static void drawRotatedRect(cv::Mat &image, const cv::RotatedRect &rotatedRect, const cv::Scalar &color, int thickness);
     static cv::Mat mainFunction(const cv::Mat &inputFrame);
     static std::vector<cv::Mat> chopFrame(const std::vector<Armor> &inputArmors, const cv::Mat &inputFrame);
